backend/requirements.txt
adtk==0.6.2
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coverage==6.1
cryptography==35.0.0
cycler==0.11.0
Django==3.2.9
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
fonttools==4.31.2
idna==3.3
ijson==3.1.4
itsdangerous==2.1.2
joblib==1.1.0
kiwisolver==1.4.2
Markdown==3.3.4
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
patsy==0.5.2
Pillow==9.1.0
psycopg2==2.9.1
py-radix==0.10.0
pycparser==2.21
PyJWT==2.3.0
pymongo==3.12.1
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
responses==0.17.0
ripe.atlas.cousteau==1.4.2
ripe.atlas.sagan==1.3.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
socketIO-client==0.7.2
sqlparse==0.4.2
statsmodels==0.13.2
tabulate==0.8.9
threadpoolctl==3.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
django-ninja==0.17.0
websocket-client==0.59.0
numpy
adtk
ijson
itsdangerous
py-radix
xgboost